pyproject.toml
[project]
name = "acoupi-batdetect2"
version = "0.1.0"
description = "An acoupi-compatible BatDetect2 model and program."
authors = [
  { name = "Aude Vuilliomenet", email = "aude.vuilliomenet.18@ucl.ac.uk" },
  { name = "Santiago Martinez Balvanera", email = "santiago.balvanera.20@ucl.uk.ac" },
]

dependencies = [
  "acoupi",
  "batdetect2>=1.0.7",
  "pydantic-settings>=2.1.0",
  "pygments>=2.18.0",
]

requires-python = ">=3.9,<3.11"
readme = "README.md"
license = { text = "Creative Commons Attribution-NonCommercial 4.0 International" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
  "pytest>=7.4.3",
  "pytest-celery>=0.0.0",
  "pyright>=1.1.363",
  "ruff>=0.4.4",
]

[tool.ruff]
line-length = 79
target-version = "py39"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 60

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "Q", "I", "D"]
ignore = ["D1"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.isort]
known-local-folder = ["src", "acoupi_batdetect2"]

[tool.coverage.run]
source = ["src/acoupi_batdetect2"]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.9"

[tool.uv.sources]
acoupi = { path = "../acoupi" }
